% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_i_squared.R
\name{bloxplot_unbiased}
\alias{bloxplot_unbiased}
\title{Plot a boxplot of the unbiased I^2 values}
\usage{
bloxplot_unbiased(plot_data)
}
\arguments{
\item{plot_data}{plotting data as generated by \code{compile_plot_data()}}
}
\value{
a ggplot2 object
}
\description{
Plot a boxplot of the unbiased I^2 values
}
