% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_i_squared.R
\name{boxplot_biased}
\alias{boxplot_biased}
\title{Plot a boxplot of the biased I^2 values}
\usage{
boxplot_biased(plot_data)
}
\arguments{
\item{plot_data}{plotting data as generated by \code{compile_plot_data()}}
}
\value{
a ggplot2 object
}
\description{
Plot a boxplot of the biased I^2 values
}
