% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_i_squared.R, R/plotting.R
\name{combine_plots}
\alias{combine_plots}
\title{Combine biased and unbiased I^2 boxplots}
\usage{
combine_plots(plot_type_1, plot_power)

combine_plots(plot_type_1, plot_power)
}
\arguments{
\item{plot_type_1}{output of \code{plot_type_1()}}

\item{plot_power}{output of \code{plot_power()}}

\item{bloxplot_unbiased}{}
}
\value{


a ggplot object saves plot as a pdf file
}
\description{
Combine biased and unbiased I^2 boxplots

Combine error rate plots
}
