% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_i_squared.R
\name{compile_plot_data}
\alias{compile_plot_data}
\title{Compile plotting data for I^2 plot}
\usage{
compile_plot_data(sim_data, scenarios)
}
\arguments{
\item{sim_data}{simulated data}

\item{scenarios}{df with scenarios}
}
\value{
a data frame suitable for the I^2 plotting
}
\description{
Compile plotting data for I^2 plot
}
