% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{compile_scenarios}
\alias{compile_scenarios}
\title{Compile a dataframe of simulation scenarios}
\usage{
compile_scenarios(
  bias_type,
  bias_percentage,
  bias_strength,
  odds_ratio,
  heterogeneity,
  ma_size,
  prob_cg_distr,
  n_cg_distr,
  bias_table
)
}
\arguments{
\item{bias_type}{Charakter vector idicating bias types}

\item{bias_percentage}{Numeric vector indicating bias percentage}

\item{bias_strength}{Character vector indicating bias strength}

\item{odds_ratio}{Numeric vector of odds ratios}

\item{heterogeneity}{Numeric vector of heterogeneities (as percentage of between study variance)}

\item{ma_size}{Numeric vector indicaing the number of individual studies per meta-analysis}
}
\value{
Returns dataframe with simulation scenarios
}
\description{
Compile a dataframe of simulation scenarios
}
