% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_functions.R
\name{compute_i_squared}
\alias{compute_i_squared}
\title{Comupte i_squared}
\usage{
compute_i_squared(or_sim, var_within, k)
}
\arguments{
\item{or_sim}{vector with odds ratios of studies in meta analysis}

\item{var_within}{vector Within study variances for each study in meta-analysis.}

\item{k}{number of studies in meta-analysis}
}
\value{
Isquared
}
\description{
Comupte i_squared
}
