% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_functions.R
\name{compute_p_value}
\alias{compute_p_value}
\title{Compute p_value.}
\usage{
compute_p_value(n, event_sim_exp, event_sim_contr)
}
\arguments{
\item{n}{size of both exposed and control group.}

\item{event_sim_exp}{Simuated number of events in exposed group.}

\item{event_sim_contr}{Simulated number of events in control-group.}
}
\value{
Returns one sided p-value from chi-square test.
}
\description{
Obtains one-sided p-value from chi-square test.
}
