% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{egger_test}
\alias{egger_test}
\title{Perform Egger test}
\usage{
egger_test(or_sim, se_lnor, sig_threshold = 0.1)
}
\arguments{
\item{or_sim}{Simulated odds ratio (as opposed to the OR give as a simulation parameter)}

\item{se_lnor}{standard error of the lnOR computed from simulated confusion matrix a,b,c,d}

\item{sig_threshold}{threshold of significace defaults to 0.1 as in the original simulation}
}
\value{
A logical value indicating whether the p value was below the sig_threshold
}
\description{
Perform Egger test
}
