% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_functions.R
\name{generate_meta_analysis}
\alias{generate_meta_analysis}
\title{Generate meta-analysis data.}
\usage{
generate_meta_analysis(
  job_id,
  scenario_id,
  bias_type,
  bias_percentage = NULL,
  bias_strength = NULL,
  odds_ratio,
  heterogeneity,
  ma_size,
  prob_cg_distr,
  n_cg_distr,
  bias_table
)
}
\arguments{
\item{job_id}{User defined job_id. Used for determining a unique seed.}

\item{scenario_id}{Unique id fer scenario = constellation of simulation parameters.}

\item{bias_type}{Bias type can either be "p" or "es".}

\item{bias_percentage}{Percentage of studies that will be removed to to publication
bias. Only needs to be provided when bias type = "es".}

\item{bias_strength}{String indicating bias severity "moderate" or "strong".
only needs to be supplied when bias type = "p".}

\item{odds_ratio}{True underlying effect of the symulated meta-analysis.}

\item{heterogeneity}{Numeric value indicating the heterogeneity fraction.}

\item{ma_size}{Intended number of studies to be included in final meta-analysis
(after publication bias).}

\item{prob_cg_distr}{Probablility of an event in the control group.
Can be any value or function that evaluates to a numeric value between 0 and 1.}

\item{n_cg_distr}{Sample size in control group (= sample size in exposed group)
Can be any value or function that evaluates to an integer.}

\item{bias_table}{List of pvalue cut-offs and corresponding selection probabilities
  list element names must correspond to elements of bias strength
#' @return A dataframe with ma-size rows. A subset of rows returned by
  simulate_unbiased_study_set()}
}
\description{
Call all helper functions to generate a dataframe with the simulated data
pertaining to a single meta-analysis.
}
