% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_functions.R
\name{get_seed}
\alias{get_seed}
\title{Get unique seed.-------------------------------------------------------------}
\usage{
get_seed(job_id, scenario_id)
}
\arguments{
\item{job_id}{User generated job_id}

\item{scenario_id}{id of the current scenario}
}
\value{
seed to be used for the combination of job_id and scenario_id
}
\description{
Get unique seed.-------------------------------------------------------------
}
