% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_functions.R
\name{obtain_true_ma_size}
\alias{obtain_true_ma_size}
\title{Obtain true number of studies that need to be simulated.}
\usage{
obtain_true_ma_size(ma_size, bias_type, bias_percentage = NULL)
}
\arguments{
\item{ma_size}{Intended number of studies after publication bias.}

\item{bias_type}{Bias type can either be "p" or "es".}

\item{bias_percentage}{Percentage of studies that will be removed to to
publication bias. Only needs to be provided when bias type = "es".}
}
\value{
Returns a numerical value indicating the adapted sample size
}
\description{
Obtain true number of studies that need to be simulated.
}
