% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{peters_test}
\alias{peters_test}
\title{Perform Peters Test}
\usage{
peters_test(or_sim, n, a, b, c, d, sig_threshold = 0.1)
}
\arguments{
\item{or_sim}{or_sim Simulated odds ratio (as opposed to the OR give as a simulation parameter)}

\item{n}{Number of aprticipants in intervention and controlgroup (i.e. half the sample size)}

\item{a}{a from confusion matrix (i.e. number of events in exposed group)}

\item{b}{b from confusion matrix (i.e. number of events in control group)}

\item{c}{c fom confusion matrix (i.e. number of non-events in exposed group)}

\item{d}{d from confusion matrix (i.e. number of non events in control group)}

\item{sig_threshold}{threshold of significace defaults to 0.1 as in the original simulation}
}
\value{
A logical value indicating whether the p value was below the sig_threshold
}
\description{
Perform Peters Test
}
