% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_error_rate.R
\name{plot_error_rate}
\alias{plot_error_rate}
\title{Plotting the error rate with various facets}
\usage{
plot_error_rate(
  df,
  titel = NULL,
  yaxis = NULL,
  facet1 = NULL,
  facet2 = NULL,
  facet3 = NULL,
  facet4 = NULL
)
}
\arguments{
\item{df}{data frame with simulation parameters and test results}

\item{titel}{optinal plot title}

\item{yaxis}{optional y-axis title}

\item{facet1}{first grouping variable (on x-axis not transformed to a factor)}

\item{facet2}{second grouping variable (transformed to factor, indicated by shape and color)}

\item{facet3}{third grouping variable (not transformed to factor facet grid groups columns)}

\item{facet4}{fourth grouping variable (not transformed to factor, facet grid rows)}
}
\value{
returns a ggplot2 object
}
\description{
Plotting the error rate with various facets
}
