% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_power}
\alias{plot_power}
\title{Plot power}
\usage{
plot_power(results_h0_false)
}
\arguments{
\item{results_h0_false}{filtered df}
}
\value{
a ggplot2 object
}
\description{
wrapper for plot_error rate for power plot
}
