% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_type_1}
\alias{plot_type_1}
\title{Plot type 1 error rate}
\usage{
plot_type_1(results_h0_true)
}
\arguments{
\item{results_h0_true}{filtered df}
}
\value{

}
\description{
wrapper for plot_error rate for type 1 error rate plot
}
