% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_functions.R
\name{set_selection_indicator}
\alias{set_selection_indicator}
\title{Set selection idicator.}
\usage{
set_selection_indicator(bias_type, p_value, bias_strength, bias_table)
}
\arguments{
\item{bias_type}{Bias type can either be "p" or "es".}

\item{p_value}{p_value (one tailed)}

\item{bias_strength}{String indicating the bias strength can be "moderate" or "severe"
passed to \code{select_prob()}}

\item{bias_table}{passed to \code{select_prob()}}
}
\value{
TRUE if bias type is "es" (selection will then be performed later).
        TRUE if bias type is "p" and the draw from the binomial distribution returned 1.
        FALSE if bias type is "p" and the draw from the binomial distribution returned 1.
}
\description{
If bias type is "p" one sample will be drawn from a binomial distribution with
the probability parameter dertermined by \code{select_prob()} based on bias strength
and size of p value.
}
