% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation_functions.R
\name{simulate_study}
\alias{simulate_study}
\title{Simulate single study.}
\usage{
simulate_study(
  job_id,
  scenario_id,
  p_contr,
  bias_type,
  bias_strength = NULL,
  odds_ratio,
  n_cg_distr,
  bias_table,
  tau_squared = 0
)
}
\arguments{
\item{job_id}{user defined job_id.}

\item{scenario_id}{id of current scenario..}

\item{p_contr}{Probability of events in control group.}

\item{bias_type}{Bias type can either be "p" or "es".}

\item{bias_strength}{String indicating bias severity "moderate" or "strong".
only needs to be supplied when bias type = "p".}

\item{odds_ratio}{True underlying effect of the symulated meta-analysis.#'}

\item{n_cg_distr}{Sample size in control group (= sample size in exposed group)
Can be any value or function that evaluates to an integer.}

\item{tau_squared}{Product of hereogeneity parameter and mean within study-variance}
}
\value{
A list of descriptives for one simulated study.
}
\description{
Simulates a single study based on the input parameters.
}
